server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
aiosqlite==0.20.0
PyJWT==2.10.1
bcrypt==4.2.1
Pillow==11.1.0
python-dotenv==1.0.1
aiohttp==3.11.11
psutil==6.1.1
